Add static linking and defining paths
The paths to the static libraries are defined by the `OPENSSL_ROOT_DIR`
and `LIBSSH2_ROOT_DIR` environment variables. These can be defined using
the `cmake -D..` optional argument. If they are not specified the static
libraries are searched in the usual places (`/usr/local/lib`). Not sure
where static libraries are searched for on Windows yet.