Change build to disable pointer sign warnings
There are several of these warnings where the `libssh2` function is clearly
dealing with binary data, such as the SFTP read/write functions. It is more
appropriate to use uint8_t for binary data than `char`. Not all binary data are
characters, signed or unsigned. The use of `uint8_t` provides better
documentation of the intention of the functions, i.e. it lets the developer know
that this function deals with binary data as opposed to strings (char*). This
however, causes a sign change.

This might be a warning and ultimately a problem, but at the moment it does not
feel like it and the warnings are annoying.