Fix authentication by Public Key
The `...fromfile` authentication does not appear to work. I believe it has
something to do with paths and finding the public/private keys, but the
`...frommemory` does work. The data is string data. So, the function signature
is changed to indicate the public/private key data is string in origin, not
binary.

I have upgraded to v1.8.2 of the libssh2 library. I have tried hard-coding paths
in this library for the public and private keys. I have tried setting `NULL` for
the public key, I have tried DOS paths and UNIX paths. I have tried relative
paths using the `~` folder, and I have tried absolute paths. The `...fromfile`
function just will not work on my Windows 10 development computer. I have not
tried libssh2 `master`, but I don't think this will help. The `...frommemory`
works just fine if I read the files outside of this library first. So, I think
it is something with paths and I am kind of done trying to figure it out.