Add API for getting known hosts
After reviewing the source code for libssh2 and some empirical testing, I can
add back in the `first` and `next` functions for iterating through the list of
known hosts. I also added some functions that get and check at the same time.

Basically, libssh2 handles the allocation and freeing of the known hosts within
the knownhosts collection. Thus, when the `libssh2_knownhost_free` function is
called, any inner nodes are freed. The wrapper structure only needs to maintain
a pointer (handle) to the inner nodes and "accessor" methods. When destroying
the wrapper structure, only the handle needs to be freed, not the internal known
host node.