Add options to configure build
Options are added to turn off automatic dependency management. These will be
useful for "offline" builds or where the developer already has the dependencies
installed on his or her system. This is mostly for UNIX-like systems and does
not really pertain to Windows. The default is to automatically handle the
dependencies.