Change linking if dependencies are not managed
If the dependencies are _not_ supposed to be built and manage, then this is an
indicator that the OpenSSL and libssh2 files are not to be statically
compiled/linked into the shared library. The assumption is these dependencies
exist on the UNIX-like system in the `/usr/lib` and the `/usr/include` folders.