Fix dynamic linking on UNIX-like systems when not building deps
The libssh2 (ssh2) and libcrypto (crypto) libraries were not being dynamically
linked and included as dependencies when the `BUILD_DEPS` variable was `OFF`.
This meant that if a shared library (SO) on NI Linux RT was build, the
dependencies would not be found and would result in an Error 13 for the Call
Library Function node in LabVIEW.

Now, the `ssh2` and `crypto` libraries are linked to the shared library if the
`BUILD_DEPS` variable is set to `OFF`. The dependencies for these shared
libraries are now listed using the `ldd` command on NI Linux RT.