Remove NI Eclipse build files
The shared library cannot be cross-compiled with NI Eclipse. The libssh2 and OpenSSL dependencies are just too hassle to get to cross-compile with Eclipse. The best way is to use a cRIO to natively compile the shared library. Thus, I am removing the .project and .cproject Eclipse files to avoid confusion and clutter.