Add function to determine file type
There was no functionality to determine the file type when obtaining the
attributes of a file on a remote server via the SFTP API. The
`lv_libssh2_sftp_attributes_file_type` function is added for the
`sftp-attributes` "module" and type. This uses some macros from the libssh2
library to determine the file type. An enum is added for convenience with
LabVIEW.