Add better SFTP error handling
SFTP error codes are all positive, while the rest of the error codes are all
negative. In the LabVIEW library, negative error codes are errors, zero is No
Error, and postive values are warnings. SFTP errors were not being properly
handled in this library. In fact, they were pretty much ignored, which meant the
LabVIEW library was not catching SFTP errors.

Better error handling is added for SFTP API. Now, if a SFTP error occurs, a
status is returned that yields a negative value so the LabVIEW library will
properly catch the error.