Add `overwrite` parameter to SFTP renaming
The `overwrite` parameter is added to the renaming function for the SFTP API.
Alberto requested that this be exposed as an option. If the value evaluates to
`TRUE`, then if a file already exists with the name, it is overwritten. If the
value exvalues to `FALSE`, i.e. zero (0), then if the file already exists with
the name, then the operation fails.