Change overwrite implementation and permissions type
The overwrite implementation for renaming files/folders with the SFTP API is
changed. During testing by Alberto, the overwrite functionality did not appear
to work. I have changed the implementation to first initialize the flags to zero
and then or values into the flags variable. This is closer to the usage of the
underlying libssh2 SFTP rename function provided in the libssh2 documentation
and examples than before, where I was initializing by OR-ing some of the
constants.

I also found out that the permissions for the libssh2 SFTP open function are
ultimately unsigned 32-bit integers even though the function uses a `long` type,
a.k.a. i32 on most systems. So, to be consistent with the permissions parameter
and type of the SFTP attributes type, I have changed all instances to u32.