Add explicit type cast
This was thought to be needed as part of an investigation and resolution for setting the permissions, a.k.a. modes, for a file. There was confusion whether permissions were stored as signed or unsigned long values within the libssh2 library. The explicit cast was added to ensure the type matched the libssh2 function signature.